% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_files}
\alias{convert_files}
\alias{convert_files_inplace}
\title{Convert files to use datasets versions of penguins and penguins_raw}
\usage{
convert_files(input, output, extensions = c("R", "r", "qmd", "rmd", "Rmd"))

convert_files_inplace(input, extensions = c("R", "r", "qmd", "rmd", "Rmd"))
}
\arguments{
\item{input}{A character vector of file paths to convert.}

\item{output}{A character vector of output file paths,
or NULL to modify files in place.
If provided, must be the same length as \code{input}.}

\item{extensions}{A character vector of file extensions to process,
defaults to c("R", "r", "qmd", "rmd", "Rmd").}
}
\value{
An invisible list with two components:
\itemize{
\item \code{changed}: A named character vector of output paths for files that
were modified, with input paths as the names.
\item \code{not_changed}: A named character vector of output paths for files
that were not modified, with input paths as names. Files are not changed
if they do not contain references to palmerpenguins (i.e. the patterns)
listed in the Details section, or if they do not have a specified
\code{extension}.
}
}
\description{
This function converts files that use the
\href{https://CRAN.R-project.org/package=palmerpenguins}{palmerpenguins} package
to use the versions of \code{penguins} and \code{penguins_raw} included in the datasets
package in R 4.5.0. It removes calls to \code{library(palmerpenguins)} and makes
necessary changes to the variable names.
}
\details{
Files are converted by:
\itemize{
\item Replacing \code{library(palmerpenguins)} with \code{""}
\item Replacing variable names:
\itemize{
\item \code{bill_length_mm} -> \code{bill_len}
\item \code{bill_depth_mm} -> \code{bill_dep}
\item \code{flipper_length_mm} -> \code{flipper_len}
\item \code{body_mass_g} -> \code{body_mass}
}
\item ends_with("\emph{mm")\verb{->}starts_with("flipper}"), starts_with(bill_)`
}

Non-convertible files (those without the specified extensions) are copied to
the output location if \code{output} is provided, but are not modified.
}
\examples{
# Get a sample file that uses palmerpenguins
penguin_file <- penguins_examples("penguins.R")

# Create a temporary file for output
output_file <- withr::local_tempfile(fileext = ".R")

# Convert the file
result <- convert_files(penguin_file, output_file)

# View the changes made
cat(readLines(output_file), sep = "\n")

# Convert multiple files to new locations
input_files <- c(
  penguins_examples("penguins.R"),
  penguins_examples("analysis/penguins.qmd")
)

# Create temporary output files
output_files <- c(
  withr::local_tempfile(fileext = ".R"),
  withr::local_tempfile(fileext = ".qmd")
)

# Convert all files
result <- convert_files(input_files, output_files)

# Modify files in-place
input_file <- penguins_examples("penguins.R")

# Copy file so don't overwrite example provided by package
copy_path <- withr::local_tempfile(fileext = ".R")
file.copy(input_file, copy_path)
convert_files_inplace(copy_path)

\dontrun{ # don't overwrite the example files provided by the package
  # Get all example files
  input_files <- penguins_examples(recursive = TRUE, full.names = TRUE)

  # Convert them in place
  convert_files_inplace(input_files)
}

}
\seealso{
\code{\link[=convert_dir]{convert_dir()}}, \code{\link[=penguins_examples]{penguins_examples()}}
}
