% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_files}
\alias{example_files}
\alias{example_dir}
\title{List or find example files from basepenguins package}
\usage{
example_files(path = NULL, full.names = FALSE, recursive = TRUE)

example_dir(copy.dir = NULL)
}
\arguments{
\item{path}{Character string. If \code{NULL} (default), lists all available example files.
If specified, returns the full path to the specified file or directory.}

\item{full.names}{Logical. If \verb{TRUE``, returns full file paths rather than relative paths. Only used when }path = NULL`. Default is FALSE.}

\item{recursive}{Logical. If \verb{TRUE``, lists files in subdirectories recursively. Only used when }path = NULL`. Default is TRUE.}

\item{copy.dir}{Character string. A directory name or path to a directory
into which the files in the example directory will be copied.
If \code{NULL} (default) returns path to example directory installed with the package.}
}
\value{
\itemize{
\item \code{example_files}:
\itemize{
\item If \code{path = NULL}, returns a character vector of available file/directory names.
\item If \code{path} is specified, returns the full file path to the requested resource.
}
\item \code{example_dir}: The path to the directory containing all examples.
}
}
\description{
These functions provides access to example files included with the basepenguins package.
When \code{example_files()} is called with \code{path = NULL}, it lists available example files.
When called with a specific path, it returns the full path to that file.
}
\examples{
# List all files and directories in the root directory
example_files()
example_files(recursive = FALSE)
example_files(full.names = TRUE)

# Get the full path to a specific example files
example_files("penguins.R")
example_files("nested/penguins.qmd")

# Get the path to all examplesss
example_dir()

}
